user-info-app-backend/requirements.txt
Flask==2.2.3
opentelemetry-api==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-exporter-otlp==1.16.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-exporter-otlp-proto-grpc==1.16.0
